requirements.txt
Django==3.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

